pyproject.toml
[project]
name = "chaos-box"
dynamic = ["version"]
description = "Collection of handy utils written in Python 3"
authors = [
    {name = "ak1ra", email = "git@ak1ra.xyz"},
]
dependencies = [
    "argcomplete>=3.6.2",
    "fastbencode>=0.3.2",
    "httpx>=0.28.1",
    "mobi>=0.3.3",
    "netaddr>=1.3.0",
    "numpy>=2.3.0",
    "opencv-python-headless>=4.11.0.86",
    "pathspec>=0.12.1",
    "Pillow>=11.2.1",
    "psutil>=7.0.0",
    "py7zr>=1.0.0",
    "pypng>=0.20220715.0",
    "pyzbar>=0.1.9",
    "qrcode>=8.2",
]
requires-python = ">=3.11"
classifiers = [
    "Environment :: Console",
]
keywords = []
readme = {file = "README.md", content-type = "text/markdown"}

# A previous PEP had specified license to be a table with a file or a text key, this format is now deprecated.
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#license-and-license-files
license = "MIT"


[project.urls]
Repository = "https://github.com/ak1ra-lab/chaos-box"
Issues = "https://github.com/ak1ra-lab/chaos-box/issues"
Changelog = "https://github.com/ak1ra-lab/chaos-box/blob/master/CHANGELOG.md"


[project.scripts]
deb-extract = "chaos_box.cmd.deb_extract:main"
archive-dirs = "chaos_box.cmd.archive_dirs:main"
merge-ip-ranges = "chaos_box.cmd.merge_ip_ranges:main"
netstats = "chaos_box.cmd.netstats:main"
pconv = "chaos_box.cmd.pconv:main"
qbt-dump = "chaos_box.cmd.qbt_dump:main"
qbt-migrate = "chaos_box.cmd.qbt_migrate:main"
qrcode-merge = "chaos_box.cmd.qrcode_merge:main"
qrcode-split = "chaos_box.cmd.qrcode_split:main"
rename-with-date = "chaos_box.cmd.rename_with_date:main"
rotate-images = "chaos_box.cmd.rotate_images:main"
shasum-list = "chaos_box.cmd.shasum_list:main"
sort-keys = "chaos_box.cmd.sort_keys:main"
urlencode = "chaos_box.cmd.urlencode:main"


[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true


[tool.pdm.version]
source = "scm"


[dependency-groups]
tests = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]
lint = [
    "ruff>=0.9.4",
]
dev = [
    {include-group = "tests"},
    {include-group = "lint"},
    "nox>=2024.10.9",
    "pre-commit>=4.2.0",
]
